requirements.txt
anthropic>=0.40.0
yfinance>=0.2.40
pandas>=2.0.0
ta>=0.11.0
python-dotenv>=1.0.0
futu-api>=9.0.0
